+++ src/output/transform.rs
@@ -294,27 +294,33 @@ fn handle_item(
             ));
         }
 
-        Item::MixinDeclaration(ref name, ref args, ref body) => scope
-            .define_mixin(
+        Item::MixinDeclaration(ref name, ref args, ref body, ref pos) => {
+            scope.define_mixin(
                 name.into(),
                 Mixin {
                     args: args.clone(),
                     scope: scope.clone(),
                     body: body.clone(),
+                    pos: pos.clone(),
                 },
-            ),
-        Item::MixinCall(ref name, ref args, ref body) => {
+            )
+        }
+        Item::MixinCall(ref name, ref args, ref body, ref pos) => {
             let sel = scope.get_selectors().clone();
             if let Some(mixin) = scope.get_mixin(&name.into()) {
                 let subscope = mixin
                     .args
                     .eval(
-                        ScopeRef::sub_selectors(mixin.scope, sel),
+                        ScopeRef::sub_selectors(mixin.scope.clone(), sel),
                         args.evaluate(scope.clone())?,
                     )
                     .map_err(|e| match e {
                         sass::ArgsError::Eval(e) => e,
-                        ae => Error::S(ae.to_string()),
+                        ae => Error::BadCall(
+                            ae.to_string(),
+                            pos.in_call(name),
+                            Some(mixin.pos.clone()),
+                        ),
                     })?;
                 subscope.define_mixin(
                     Name::from_static("%%BODY%%"),
@@ -322,6 +328,8 @@ fn handle_item(
                         args: FormalArgs::none(),
                         scope,
                         body: body.clone().unwrap_or_else(Mixin::no_body),
+                        // TODO: Get a better pos?
+                        pos: pos.clone(),
                     },
                 );
                 handle_body(
@@ -331,12 +339,27 @@ fn handle_item(
                     buf,
                     subscope,
                     file_context,
-                )?;
+                )
+                .map_err(|e: Error| match e {
+                    Error::BadArguments(msg, decl) => {
+                        let pos = pos.in_call(name);
+                        Error::BadCall(msg, pos, Some(decl))
+                    }
+                    Error::AtError(msg, _pos) => {
+                        let msg = format!("Error: {}", msg);
+                        Error::BadCall(msg, pos.clone(), None)
+                    }
+                    e => {
+                        let pos = pos.in_call(name);
+                        Error::BadCall(e.to_string(), pos, None)
+                    }
+                })?;
             } else {
-                return Err(Error::S(format!(
-                    "Unknown mixin {}({:?})",
-                    name, args
-                )));
+                return Err(Error::BadCall(
+                    format!("Unknown mixin {}", name),
+                    pos.clone(),
+                    None,
+                ));
             }
         }
         Item::Content => {